requirements.txt
streamlit>=1.49.0
google-generativeai>=0.8.0
typing-extensions>=4.0.0
